refactor: 전체 적용 쿠폰 선택 로직 Hook으로 분리
