+++ src/components/CartPage/CouponSelect/AllCouponSelect.tsx
@@ -1,30 +1,9 @@
-import { useRecoilState, useRecoilValue } from 'recoil';
-import { couponsSelector, selectedCouponsState } from '../../../atoms/coupons';
 import * as S from './AllCouponSelect.styles';
-import { selectedCartItemIdsState } from '../../../atoms/cart';
-import { useEffect } from 'react';
+import useAllCouponSelect from '../../../hooks/useAllCouponSelect';
 
 const AllCouponSelect = () => {
-  const couponsState = useRecoilValue(couponsSelector);
-  const [selectedCoupons, setSelectedCoupons] =
-    useRecoilState(selectedCouponsState);
-  const selectedCartIds = useRecoilValue(selectedCartItemIdsState);
-
-  useEffect(() => {
-    if (selectedCartIds.size === 0) {
-      setSelectedCoupons([]);
-    }
-  }, [selectedCartIds, setSelectedCoupons]);
-
-  const onSelectedCouponsChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    if (e.target.value === 'DEFAULT' || selectedCartIds.size === 0) {
-      e.target.selectedIndex = 0;
-      setSelectedCoupons([]);
-      return;
-    }
-    const couponId = +e.target.value;
-    setSelectedCoupons([couponId]);
-  };
+  const { couponsState, selectedCoupons, onSelectedCouponsChange } =
+    useAllCouponSelect();
   return (
     <S.Select onChange={onSelectedCouponsChange}>
       <option defaultChecked value={'DEFAULT'}>
